Mark RPC implementation as moved into rpc-methods folder
